Change release task to reset the database
This task used to be run on each deployment, but now that the
deploy section from fly.toml has been removed, I don't think it
does any longer, which means this task should only be run when called
manually.

The last time that I tried to deploy the application with the deploy section
still in fly.toml, the release task failed anyway with the error
noted in the comment.